Moved docs_definition.py and updated files
